Correct `Send` and `Sync` bounds
This is necessary for thread safety, as otherwise you could, e.g., put a `Rc<Cell<T>>` in here and send it across threads (which is not safe in this case).  I believe, however, that `Sync` only requires `Send`, not `Sync`, as the ring buffer already performs adequate synchronization during `get` and you do not hand out shared references to the items in the queue.